Sharp has been added

+++ controllers/book.js
@@ -5,31 +5,84 @@ const sharp = require("sharp");
 exports.getAllBooks = (req, res, next) => {
   Book.find()
     .then((books) => {
+      if (books.length === 0) {
+        return res.status(404).json({
+          message: "Aucun livre existant",
+        });
+      }
       res.status(200).json(books);
     })
-    .catch((error) => res.status(404).json({ error }));
+    .catch((error) => res.status(500).json({ error }));
 };
 
 exports.getOneBook = (req, res, next) => {
   Book.findOne({ _id: req.params.id })
-    .then((book) => res.status(200).json(book))
-    .catch((error) => res.status(404).json({ error }));
+    .then((book) => {
+      res.status(200).json(book);
+    })
+    .catch((error) => {
+      if (error.name === "CastError" && error.kind === "ObjectId") {
+        return res.status(404).json({
+          message: "Le livre que vous recherchez n'existe pas.",
+        });
+      } 
+      
+      res.status(500).json({ error });
+     
+    });
 };
 
 exports.getBestRatedBooks = (req, res, next) => {
   Book.find()
     .sort({ averageRating: -1 })
     .limit(3)
-    .then((bestRatesBooks) => {
-      res.status(200).json(bestRatesBooks);
+    .then((bestRatedBooks) => {
+      if (bestRatedBooks.length === 0) {
+        return res.status(404).json({
+          message: "Aucun livre existant",
+        });
+      }
+      res.status(200).json(bestRatedBooks);
     })
-    .catch((error) => res.status(404).json({ error }));
+    .catch((error) => res.status(500).json({ error }));
 };
 
 exports.createBook = (req, res, next) => {
   const bookObject = JSON.parse(req.body.book);
   delete bookObject._id;
   delete bookObject.userId;
+
+  if (!bookObject.title) {
+    return res.status(400).json({ message: "Veuillez saisir un titre" });
+  }
+
+  if (!bookObject.author) {
+    return res.status(400).json({ message: "Veuillez saisir un nom d'auteur" });
+  }
+
+  if (!bookObject.year) {
+    return res
+      .status(400)
+      .json({ message: "Veuillez saisir une date de parution" });
+  }
+
+  if (!bookObject.genre) {
+    return res.status(400).json({ message: "Veuillez saisir un genre" });
+  }
+
+  if (!bookObject.ratings[0].grade) {
+    return res.status(400).json({ message: "Veuillez saisir une note" });
+  }
+
+  if (!bookObject.averageRating) {
+    return res.status(400).json({ message: "Veuillez saisir une note" });
+  }
+
+  if (!req.file) {
+    return res.status(400).json({ message: "Veuillez ajouter une image" });
+  }
+
+
 
   const book = new Book({
     ...bookObject,
@@ -41,7 +94,9 @@ exports.createBook = (req, res, next) => {
 
   book
     .save()
-    .then(() => res.status(201).json({ message: "Livre créé" }))
+    .then(() => {
+      res.status(201).json({ message: "Livre ajouté" });
+    })
     .catch((error) => res.status(500).json({ error }));
 };
 
@@ -62,7 +117,9 @@ exports.rateBook = (req, res, next) => {
         )
           .then((updatedBook) => {
             if (!updatedBook) {
-              return res.status(404).json({ message: "Livre non trouvé." });
+              return res
+                .status(404)
+                .json({ message: "Le livre n'a pas pu être récupéré" });
             }
             const totalRatings = updatedBook.ratings.length;
             const sumOfRatings = updatedBook.ratings.reduce(
@@ -76,6 +133,11 @@ exports.rateBook = (req, res, next) => {
             updatedBook
               .save()
               .then(() => {
+                if (!updatedBook) {
+                  return res
+                    .status(404)
+                    .json({ message: "Le livre n'a pas pu être mis à jour" });
+                }
                 res.status(200).json(updatedBook);
               })
               .catch((error) => {
@@ -88,6 +150,11 @@ exports.rateBook = (req, res, next) => {
       }
     })
     .catch((error) => {
+      if (error.name === "CastError" && error.kind === "ObjectId") {
+        return res.status(404).json({
+          message: "Le livre que vous essayer de noter n'existe pas.",
+        });
+      } 
       res.status(500).json({ error });
     });
 };
@@ -103,10 +170,38 @@ exports.updateBook = (req, res, next) => {
     : { ...req.body };
 
   delete bookObject.userId;
+
+
+  if (!bookObject.title) {
+    return res.status(400).json({ message: "Veuillez saisir un titre" });
+  }
+
+  if (!bookObject.author) {
+    return res.status(400).json({ message: "Veuillez saisir un nom d'auteur" });
+  }
+
+  if (!bookObject.year) {
+    return res
+      .status(400)
+      .json({ message: "Veuillez saisir une date de parution" });
+  }
+
+  if (!bookObject.genre) {
+    return res.status(400).json({ message: "Veuillez saisir un genre" });
+  }
+
+  if (!bookObject.ratings[0].grade) {
+    return res.status(400).json({ message: "Veuillez saisir une note" });
+  }
+
+  if (!bookObject.averageRating) {
+    return res.status(400).json({ message: "Veuillez saisir une note" });
+  }
+
   Book.findOne({ _id: req.params.id })
     .then((book) => {
       if (book.userId != req.auth.userId) {
-        res.status(403).json({ message: "Non-autorisé" });
+        return res.status(403).json({ message: "Modification non-autorisé" });
       } else {
         const previousImage = book.imageUrl.split("/images/")[1];
         Book.updateOne(
@@ -114,37 +209,67 @@ exports.updateBook = (req, res, next) => {
           { ...bookObject, _id: req.params.id }
         )
           .then(() => {
+            if (!book) {
+              return res
+                .status(404)
+                .json({ message: "Le livre n'a pas pu être mis à jour" });
+            }
             res.status(200).json({ message: "Livre modifié" });
-            fs.unlink("images/" + previousImage, (error) => {
-              if (error) {
-                console.log(
-                  "Une erreur s'est produite lors de la suppression de l'image: ",
-                  error
-                );
-              } else {
-                console.log("L'image a bien été supprimée");
-              }
-            });
+            if (req.file) {
+              fs.unlink("images/" + previousImage, (error) => {
+                if (error) {
+                  console.log(
+                    "Une erreur s'est produite lors de la suppression de l'image: ",
+                    error
+                  );
+                } else {
+                  console.log("L'ancienne image a bien été supprimée");
+                }
+              });
+            }
           })
           .catch((error) => res.status(500).json({ error }));
       }
     })
-    .catch((error) => res.status(500).json({ error }));
+    .catch((error) =>{ 
+      if (error.name === "CastError" && error.kind === "ObjectId") {
+        return res.status(404).json({
+          message: "Le livre que vous essayer de modifier n'existe pas.",
+        });
+      } 
+      res.status(500).json({ error })});
 };
 
 exports.deleteBook = (req, res, next) => {
   Book.findOne({ _id: req.params.id })
     .then((book) => {
-      if (book.userId != req.auth.userId) {
-        res.status(403).json({ message: "Non-autorisé" });
-      } else {
-        const filename = book.imageUrl.split("/images/")[1];
-        fs.unlink(`images/${filename}`, () => {
-          Book.deleteOne({ _id: req.params.id })
-            .then(() => res.status(200).json({ message: "Livre supprimé" }))
-            .catch((error) => res.status(404).json({ error }));
+      if (!book) {
+        return res.status(404).json({
+          message:
+            "Échec de la suppression : le livre que vous essayez de supprimer n'existe pas.",
         });
+      } else {
+        if (book.userId != req.auth.userId) {
+          res.status(403).json({ message: "Suppression non-autorisé" });
+        } else {
+          const filename = book.imageUrl.split("/images/")[1];
+          console.log("iciiii", filename);
+          fs.unlink(`images/${filename}`, () => {
+            Book.deleteOne({ _id: req.params.id })
+              .then(() => res.status(200).json({ message: "Livre supprimé" }))
+              .catch((error) => res.status(404).json({ error }));
+          });
+        }
       }
     })
-    .catch((error) => res.status(500).json({ error }));
+    .catch((error) => {
+      if (error.name === "CastError" && error.kind === "ObjectId") {
+        return res.status(404).json({
+          message:
+            "Échec de la suppression : le livre que vous essayez de supprimer n'existe pas.",
+        });
+      } else {
+        return res.status(500).json({ error });
+      }
+    });
 };
